pyproject.toml
[project]
name = "genesis-humanoid-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "genesis-world>=0.2.1",
    "matplotlib>=3.10.3",
    "mss>=10.0.0",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "torch>=2.7.1",
    "trimesh>=4.6.12",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]
